Moved `download_torrent` and `fetch` function into `functions.py` file.
