+++ src/config.py
@@ -1,5 +1,5 @@
-import os
 import json
+import os
 from logger import Logger
